refactor: tracktime Firestore (WIP)
