+++ src/utils/trackTime.ts
@@ -1,18 +1,20 @@
-import {
-  doc,
-  updateDoc,
-  getDoc,
-  setDoc,
-  increment,
-  serverTimestamp,
-} from "firebase/firestore";
-import { getFaviconUrl } from "./tabs";
+// import {
+//   doc,
+//   updateDoc,
+//   getDoc,
+//   setDoc,
+//   increment,
+//   serverTimestamp,
+// } from "firebase/firestore";
+// import { getFaviconUrl } from "./tabs";
 
-import debounce from "lodash.debounce";
+// import debounce from "lodash.debounce";
 
 import { subDays, addDays, format } from "date-fns";
 
-import { db } from "../../firebase-config";
+// import { db } from "../../firebase-config";
+
+import { firebaseService } from "./firebaseService";
 
 import { getUserId } from "../background";
 
@@ -45,7 +47,7 @@ export function trackTabTime(url: string): void {
   };
 }
 
-export async function updateTabDuration(newUrl?: string): Promise<void> {
+export async function updateUrlDuration(newUrl?: string): Promise<void> {
   if (!newUrl) return;
   const userId = await getUserId();
   const today = getCurrentDate();
@@ -54,67 +56,64 @@ export async function updateTabDuration(newUrl?: string): Promise<void> {
   if (!tabInfo) {
     return;
   }
-  const { startTime, date } = tabInfo;
+  const { startTime, date, durationBySecond } = tabInfo;
   const duration = Date.now() - startTime;
   const seconds = Math.floor(duration / 1000);
   if (seconds < 1) return;
-  tabInfo.durationBySecond += seconds;
-  if (!userId) return;
-  try {
-    const domain = new URL(newUrl).hostname;
-    const myDomain: string = "icdbgchingbnboklhnagfckgjpdfjfeg";
-    if (domain === myDomain || domain === "newtab") return;
-    const debouncedWriteToFirestore = getDebouncedWrite(userId, domain);
-    debouncedWriteToFirestore(tabInfo.durationBySecond, newUrl, date);
-  } catch (error) {
-    console.error("Error updating tab duration: ", error);
-  }
+  const newDuration = durationBySecond + seconds;
+
+  // const domain = new URL(newUrl).hostname;
+  // const myDomain: string = "icdbgchingbnboklhnagfckgjpdfjfeg";
+  // if (domain === myDomain || domain === "newtab") return;
+  // const debouncedWriteToFirestore = getDebouncedWrite(userId, domain);
+  // debouncedWriteToFirestore(tabInfo.durationBySecond, newUrl, date);
+  return firebaseService.saveUrlDuration(userId, newUrl, date, newDuration);
 }
 
-type DebouncedFunction = (
-  durationBySecond: number,
-  url: string,
-  date: string,
-) => void;
+// type DebouncedFunction = (
+//   durationBySecond: number,
+//   url: string,
+//   date: string,
+// ) => void;
 
-const debouncedWrites: Record<string, DebouncedFunction> = {};
+// const debouncedWrites: Record<string, DebouncedFunction> = {};
 
-function getDebouncedWrite(userId: string, domain: string): DebouncedFunction {
-  if (!debouncedWrites[domain]) {
-    debouncedWrites[domain] = debounce(async (durationBySecond, url, date) => {
-      const urlRef = doc(
-        db,
-        "users",
-        userId,
-        "urlDurations",
-        date,
-        "domains",
-        domain,
-      );
-      const urlSnapShot = await getDoc(urlRef);
-      if (urlSnapShot.exists()) {
-        await updateDoc(urlRef, {
-          durationBySecond: increment(durationBySecond),
-          visitCounts: increment(1),
-          lastVisit: serverTimestamp(),
-        });
-      } else {
-        await setDoc(
-          urlRef,
-          {
-            faviconUrl: getFaviconUrl(url),
-            url,
-            durationBySecond,
-            visitCounts: 1,
-            lastVisit: serverTimestamp(),
-          },
-          { merge: true },
-        );
-      }
-    }, 1000);
-  }
-  return debouncedWrites[domain];
-}
+// function getDebouncedWrite(userId: string, domain: string): DebouncedFunction {
+//   if (!debouncedWrites[domain]) {
+//     debouncedWrites[domain] = debounce(async (durationBySecond, url, date) => {
+//       const urlRef = doc(
+//         db,
+//         "users",
+//         userId,
+//         "urlDurations",
+//         date,
+//         "domains",
+//         domain,
+//       );
+//       const urlSnapShot = await getDoc(urlRef);
+//       if (urlSnapShot.exists()) {
+//         await updateDoc(urlRef, {
+//           durationBySecond: increment(durationBySecond),
+//           visitCounts: increment(1),
+//           lastVisit: serverTimestamp(),
+//         });
+//       } else {
+//         await setDoc(
+//           urlRef,
+//           {
+//             faviconUrl: getFaviconUrl(url),
+//             url,
+//             durationBySecond,
+//             visitCounts: 1,
+//             lastVisit: serverTimestamp(),
+//           },
+//           { merge: true },
+//         );
+//       }
+//     }, 1000);
+//   }
+//   return debouncedWrites[domain];
+// }
 
 export function getPrevDate(date: string): string {
   const today = new Date(date);
